requirements.txt
flask==3.1.0
flask_cors==5.0.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.43
langchain_openai==0.3.8
openai==0.27.8
python-dotenv==1.0.0
pypdf==5.3.1
PyYAML==6.0.1
chromadb==0.6.3
numpy==1.26.3
requests==2.31.0